requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dialogflow==0.6.0
Flask==1.0.2
Flask-Cors==3.0.7
futures==3.1.1
google-api-core==1.11.1
google-assistant-grpc==0.2.1
google-assistant-sdk==0.5.1
google-auth==1.6.3
google-auth-oauthlib==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
pathlib2==2.3.3
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pyOpenSSL==19.0.0
python-dotenv==0.10.1
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sounddevice==0.3.13
tenacity==4.12.0
urllib3==1.24.2
Werkzeug==0.15.2
